+++ src/StackExchange.Redis/ConfigurationOptions.cs
@@ -111,7 +111,8 @@ namespace StackExchange.Redis
                 Tunnel = "tunnel",
                 SetClientLibrary = "setlib",
                 Protocol = "protocol",
-                HighIntegrity = "highIntegrity";
+                HighIntegrity = "highIntegrity",
+                TcpKeepAlive = "tcpKeepAlive";
 
             private static readonly Dictionary<string, string> normalizedOptions = new[]
             {
@@ -132,6 +133,7 @@ namespace StackExchange.Redis
                 PreserveAsyncOrder,
                 Proxy,
                 ResolveDns,
+                ResponseTimeout,
                 ServiceName,
                 Ssl,
                 SslHost,
@@ -141,8 +143,11 @@ namespace StackExchange.Redis
                 Version,
                 WriteBuffer,
                 CheckCertificateRevocation,
+                Tunnel,
+                SetClientLibrary,
                 Protocol,
                 HighIntegrity,
+                TcpKeepAlive,
             }.ToDictionary(x => x, StringComparer.OrdinalIgnoreCase);
 
             public static string TryNormalize(string value)
@@ -169,6 +174,7 @@ namespace StackExchange.Redis
         private Version? defaultVersion;
 
         private int? keepAlive, asyncTimeout, syncTimeout, connectTimeout, responseTimeout, connectRetry, configCheckSeconds;
+        private bool? tcpKeepAlive;
 
         private Proxy? proxy;
 
@@ -598,6 +604,15 @@ namespace StackExchange.Redis
         }
 
         /// <summary>
+        /// Gets or sets whether to enable TCP keep-alive when appropriate (endpoint- and platform-dependent).
+        /// </summary>
+        public bool TcpKeepAlive
+        {
+            get => tcpKeepAlive ?? Defaults.TcpKeepAlive;
+            set => tcpKeepAlive = value;
+        }
+
+        /// <summary>
         /// The <see cref="ILoggerFactory"/> to get loggers for connection events.
         /// Note: changes here only affect <see cref="ConnectionMultiplexer"/>s created after.
         /// </summary>
@@ -851,6 +866,7 @@ namespace StackExchange.Redis
             heartbeatInterval = heartbeatInterval,
             heartbeatConsistencyChecks = heartbeatConsistencyChecks,
             highIntegrity = highIntegrity,
+            tcpKeepAlive = tcpKeepAlive,
             WriteMode = WriteMode,
 #if DEBUG
             OutputLog = OutputLog,
@@ -937,6 +953,7 @@ namespace StackExchange.Redis
             Append(sb, OptionKeys.SetClientLibrary, setClientLibrary);
             Append(sb, OptionKeys.HighIntegrity, highIntegrity);
             Append(sb, OptionKeys.Protocol, FormatProtocol(_protocol));
+            Append(sb, OptionKeys.TcpKeepAlive, tcpKeepAlive);
             if (Tunnel is { IsInbuilt: true } tunnel)
             {
                 Append(sb, OptionKeys.Tunnel, tunnel.ToString());
@@ -1104,6 +1121,9 @@ namespace StackExchange.Redis
                             break;
                         case OptionKeys.HighIntegrity:
                             HighIntegrity = OptionKeys.ParseBoolean(key, value);
+                            break;
+                        case OptionKeys.TcpKeepAlive:
+                            TcpKeepAlive = OptionKeys.ParseBoolean(key, value);
                             break;
                         case OptionKeys.Tunnel:
                             if (value.IsNullOrWhiteSpace())