+++ toys/StackExchange.Redis.Server/RedisClient.Output.cs
@@ -178,6 +178,9 @@ public partial class RedisClient
                 case RespPrefix.Integer:
                     MessageWriter.WriteInteger(output, (long)value.AsRedisValue());
                     break;
+                case RespPrefix.Double:
+                    WriteDouble(output, (double)value.AsRedisValue());
+                    break;
                 case RespPrefix.SimpleError:
                     prefix = '-';
                     goto BasicMessage;
@@ -228,6 +231,18 @@ public partial class RedisClient
                         "Unexpected result type: " + value.Type);
             }
         }
+
+        static void WriteDouble(IBufferWriter<byte> output, double value)
+        {
+            Span<byte> valueSpan = stackalloc byte[Format.MaxDoubleTextLen];
+            var len = Format.FormatDouble(value, valueSpan);
+            var span = output.GetSpan(3 + len);
+            span[0] = (byte)',';
+            valueSpan.Slice(0, len).CopyTo(span.Slice(1));
+            span[1 + len] = (byte)'\r';
+            span[2 + len] = (byte)'\n';
+            output.Advance(3 + len);
+        }
     }
 
     public RespPrefix ApplyProtocol(RespPrefix type) => IsResp2 ? ToResp2(type) : type;
@@ -239,6 +254,7 @@ public partial class RedisClient
             case RespPrefix.Boolean:
                 return RespPrefix.Integer;
             case RespPrefix.Double:
+                return RespPrefix.BulkString;
             case RespPrefix.BigInteger:
                 return RespPrefix.SimpleString;
             case RespPrefix.BulkError: